users/requirements.txt
fastapi
uvicorn
pymysql
pydantic
PyMuPDF
datetime
fitz
requests
markdownify
passlib
langchain
langchain-community
cryptography
sentence-transformers
transformers 
python-jose
Chroma
chromadb
frontend
tools
nltk
torch